fix(speaker): update design of speaker list
