+++ src/app/pages/speaker-detail/speaker-detail.html
@@ -24,4 +24,25 @@
   </div>
 
   <p>{{speaker?.about}}</p>
+
+  <ion-row class="ion-no-padding ion-justify-content-center">
+    <ion-col size="4" class="ion-text-left">
+      <ion-button fill="clear" size="small" color="primary" (click)="goToSpeakerTwitter(speaker)">
+        <ion-icon name="logo-twitter" slot="start"></ion-icon>
+        Tweet
+      </ion-button>
+    </ion-col>
+    <ion-col size="4" class="ion-text-center">
+      <ion-button fill="clear" size="small" color="primary" (click)="openSpeakerShare(speaker)">
+        <ion-icon name="share-social" slot="start"></ion-icon>
+        Share
+      </ion-button>
+    </ion-col>
+    <ion-col size="4" class="ion-text-right">
+      <ion-button fill="clear" size="small" color="primary" (click)="openContact(speaker)">
+        <ion-icon name="chatbubbles-outline" slot="start"></ion-icon>
+        Contact
+      </ion-button>
+    </ion-col>
+  </ion-row>
 </ion-content>
