+++ src/app/pages/speaker-detail/speaker-detail.ts
@@ -1,6 +1,8 @@
-import { Component, ViewEncapsulation } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { Component } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import { ConferenceData } from '../../providers/conference-data';
+import { ActionSheetController } from '@ionic/angular';
+import { InAppBrowser } from '@ionic-native/in-app-browser/ngx';
 
 @Component({
   selector: 'page-speaker-detail',
@@ -12,8 +14,10 @@ export class SpeakerDetailPage {
 
   constructor(
     private dataProvider: ConferenceData,
-    private router: Router,
-    private route: ActivatedRoute
+    private route: ActivatedRoute,
+    public actionSheetCtrl: ActionSheetController,
+    public confData: ConferenceData,
+    public inAppBrowser: InAppBrowser,
   ) {}
 
   ionViewWillEnter() {
@@ -28,5 +32,71 @@ export class SpeakerDetailPage {
         }
       }
     });
+  }
+
+  goToSpeakerTwitter(speaker: any) {
+    this.inAppBrowser.create(
+      `https://twitter.com/${speaker.twitter}`,
+      '_blank'
+    );
+  }
+
+  async openSpeakerShare(speaker: any) {
+    const actionSheet = await this.actionSheetCtrl.create({
+      header: 'Share ' + speaker.name,
+      buttons: [
+        {
+          text: 'Copy Link',
+          handler: () => {
+            console.log(
+              'Copy link clicked on https://twitter.com/' + speaker.twitter
+            );
+            if (
+              (window as any)['cordova'] &&
+              (window as any)['cordova'].plugins.clipboard
+            ) {
+              (window as any)['cordova'].plugins.clipboard.copy(
+                'https://twitter.com/' + speaker.twitter
+              );
+            }
+          }
+        },
+        {
+          text: 'Share via ...'
+        },
+        {
+          text: 'Cancel',
+          role: 'cancel'
+        }
+      ]
+    });
+
+    await actionSheet.present();
+  }
+
+  async openContact(speaker: any) {
+    const mode = 'ios'; // this.config.get('mode');
+
+    const actionSheet = await this.actionSheetCtrl.create({
+      header: 'Contact ' + speaker.name,
+      buttons: [
+        {
+          text: `Email ( ${speaker.email} )`,
+          icon: mode !== 'ios' ? 'mail' : null,
+          handler: () => {
+            window.open('mailto:' + speaker.email);
+          }
+        },
+        {
+          text: `Call ( ${speaker.phone} )`,
+          icon: mode !== 'ios' ? 'call' : null,
+          handler: () => {
+            window.open('tel:' + speaker.phone);
+          }
+        }
+      ]
+    });
+
+    await actionSheet.present();
   }
 }
