+++ src/app/pages/speaker-list/speaker-list.scss
@@ -3,21 +3,37 @@
 }
 
 .speaker-card {
-  height: 100%;
   display: flex;
   flex-direction: column;
+}
+
+.speaker-card .speaker-item {
+  --min-height: 85px;
+}
+
+.speaker-card .speaker-item h2 {
+  font-size: 18px;
+  font-weight: 500;
+  letter-spacing: 0.02em;
+}
+
+.speaker-card .speaker-item p {
+  font-size: 13px;
+  letter-spacing: 0.02em;
 }
 
 .speaker-card ion-card-header {
   padding: 0;
 }
 
-.speaker-card ion-card-header .item {
-  padding: 4px 16px;
-}
-
 .speaker-card ion-card-content {
   flex: 1 1 auto;
+
+  padding: 0;
+}
+
+.md ion-list {
+  border-top: 1px solid var(--ion-color-step-150, #d7d8da);
 
   padding: 0;
 }
