+++ src/assets/data/data.json
@@ -250,6 +250,7 @@
       "profilePic": "assets/img/speakers/bear.jpg",
       "twitter": "ionicframework",
       "about": "Burt is a Bear.",
+      "title": "Software Engineer",
       "location": "Everywhere",
       "email": "burt@example.com",
       "phone": "+1-541-754-3010",
@@ -260,6 +261,7 @@
       "profilePic": "assets/img/speakers/cheetah.jpg",
       "twitter": "ionicframework",
       "about": "Charlie is a Cheetah.",
+      "title": "Software Engineer",
       "location": "Everywhere",
       "email": "charlie@example.com",
       "phone": "+1-541-754-3010",
@@ -270,6 +272,7 @@
       "profilePic": "assets/img/speakers/duck.jpg",
       "twitter": "ionicframework",
       "about": "Donald is a Duck.",
+      "title": "Software Engineer",
       "location": "Everywhere",
       "email": "donald@example.com",
       "phone": "+1-541-754-3010",
@@ -280,6 +283,7 @@
       "profilePic": "assets/img/speakers/eagle.jpg",
       "twitter": "ionicframework",
       "about": "Eva is an Eagle.",
+      "title": "Developer Advocate",
       "location": "Everywhere",
       "email": "eva@example.com",
       "phone": "+1-541-754-3010",
@@ -290,6 +294,7 @@
       "profilePic": "assets/img/speakers/elephant.jpg",
       "twitter": "ionicframework",
       "about": "Ellie is an Elephant.",
+      "title": "Software Engineer",
       "location": "Everywhere",
       "email": "ellie@example.com",
       "phone": "+1-541-754-3010",
@@ -300,6 +305,7 @@
       "profilePic": "assets/img/speakers/giraffe.jpg",
       "twitter": "ionicframework",
       "about": "Gino is a Giraffe.",
+      "title": "Software Engineer",
       "location": "Everywhere",
       "email": "gino@example.com",
       "phone": "+1-541-754-3010",
@@ -310,6 +316,7 @@
       "profilePic": "assets/img/speakers/iguana.jpg",
       "twitter": "ionicframework",
       "about": "Isabella is an Iguana.",
+      "title": "Software Engineer",
       "location": "Everywhere",
       "email": "isabella@example.com",
       "phone": "+1-541-754-3010",
@@ -320,6 +327,7 @@
       "profilePic": "assets/img/speakers/kitten.jpg",
       "twitter": "ionicframework",
       "about": "Karl is a Kitten.",
+      "title": "Developer Advocate",
       "location": "Everywhere",
       "email": "karl@example.com",
       "phone": "+1-541-754-3010",
@@ -330,6 +338,7 @@
       "profilePic": "assets/img/speakers/lion.jpg",
       "twitter": "ionicframework",
       "about": "Lionel is a Lion.",
+      "title": "Developer Advocate",
       "location": "Everywhere",
       "email": "lionel@example.com",
       "phone": "+1-541-754-3010",
@@ -340,6 +349,7 @@
       "profilePic": "assets/img/speakers/mouse.jpg",
       "twitter": "ionicframework",
       "about": "Molly is a Mouse.",
+      "title": "Software Engineer",
       "location": "Everywhere",
       "email": "molly@example.com",
       "phone": "+1-541-754-3010",
@@ -350,6 +360,7 @@
       "profilePic": "assets/img/speakers/puppy.jpg",
       "twitter": "ionicframework",
       "about": "Paul is a Puppy.",
+      "title": "Software Engineer",
       "location": "Everywhere",
       "email": "paul@example.com",
       "phone": "+1-541-754-3010",
@@ -360,6 +371,7 @@
       "profilePic": "assets/img/speakers/rabbit.jpg",
       "twitter": "ionicframework",
       "about": "Rachel is a Rabbit.",
+      "title": "Senior Software Engineer",
       "location": "Everywhere",
       "email": "rachel@example.com",
       "phone": "+1-541-754-3010",
@@ -370,6 +382,7 @@
       "profilePic": "assets/img/speakers/turtle.jpg",
       "twitter": "ionicframework",
       "about": "Ted is a Turtle.",
+      "title": "Software Engineer",
       "location": "Everywhere",
       "email": "ted@example.com",
       "phone": "+1-541-754-3010",
